+++ modules/data_structures/builtin/include/namedargs2cellBuiltin.hpp
@@ -23,38 +23,17 @@
 // License along with this program. If not, see <http://www.gnu.org/licenses/>.
 // LICENCE_BLOCK_END
 //=============================================================================
-#include "NelsonGateway.hpp"
-#include "structBuiltin.hpp"
-#include "iscellstrBuiltin.hpp"
-#include "cellBuiltin.hpp"
-#include "fieldnamesBuiltin.hpp"
-#include "struct2cellBuiltin.hpp"
-#include "cell2structBuiltin.hpp"
-#include "cellfunBuiltin.hpp"
-#include "isfieldBuiltin.hpp"
-#include "namedargs2cellBuiltin.hpp"
+#pragma once
 //=============================================================================
-using namespace Nelson;
+#include "ArrayOf.hpp"
+#include "Evaluator.hpp"
 //=============================================================================
-const std::wstring gatewayName = L"data_structures";
+namespace Nelson {
 //=============================================================================
-static const nlsGateway gateway[] = {
-    { "struct", Nelson::DataStructuresGateway::structBuiltin, 1, 1 },
-    { "iscellstr", Nelson::DataStructuresGateway::iscellstrBuiltin, 1, 1 },
-    { "cell", Nelson::DataStructuresGateway::cellBuiltin, 1, 0 },
-    { "fieldnames", Nelson::DataStructuresGateway::fieldnamesBuiltin, 1, 1 },
-    { "struct2cell", Nelson::DataStructuresGateway::struct2cellBuiltin, 1, 1 },
-    { "cell2struct", Nelson::DataStructuresGateway::cell2structBuiltin, 1, 3 },
-    { "cellfun", Nelson::DataStructuresGateway::cellfunBuiltin, -1, -1 },
-    { "isfield", Nelson::DataStructuresGateway::isfieldBuiltin, 1, 2 },
-    { "namedargs2cell", Nelson::DataStructuresGateway::namedargs2cellBuiltin, 1, 1 },
-};
+namespace DataStructuresGateway {
+    ArrayOfVector
+    namedargs2cellBuiltin(Evaluator* eval, int nLhs, const ArrayOfVector& argIn);
+}
 //=============================================================================
-NLSGATEWAYFUNC(gateway)
-//=============================================================================
-NLSGATEWAYINFO(gateway)
-//=============================================================================
-NLSGATEWAYREMOVE(gateway)
-//=============================================================================
-NLSGATEWAYNAME()
+} // namespace Nelson
 //=============================================================================